--- src/components/ColorPicker.scss
@@ -1,4 +1,0 @@
-.ColorPicker {
-  max-width: 40px;
-  max-height: 40px;
-}
