--- src/components/ColorPicker.tsx
@@ -1,14 +1,0 @@
-import React from 'react';
-
-// const colors = ['red', 'green', 'blue'];
-
-export const ColorPicker: React.FC = () => {
-  return (
-    <button
-      type="button"
-      className=""
-    >
-      Hello!
-    </button>
-  );
-};
